+++ src/Var.jl
@@ -14,6 +14,7 @@ import ..Utils:
     seconds_to_prettystr,
     squeeze,
     split_by_season,
+    split_by_season_across_time,
     time_to_date,
     date_to_time,
     _data_at_dim_vals,
@@ -48,6 +49,7 @@ export OutputVar,
     integrate_lat,
     isempty,
     split_by_season,
+    split_by_season_across_time,
     bias,
     global_bias,
     squared_error,
@@ -1306,6 +1308,35 @@ function split_by_season(var::OutputVar)
 
     return _split_along_dim(var, time_name(var), season_times)
 end
+
+"""
+    split_by_season_across_time(var::OutputVar)
+
+Return a vector of `OutputVar`s split by season across time.
+
+The months of the seasons are March to May, June to August, September to November, and
+December to February. If there are no dates found for a season, then the `OutputVar` for
+that season will be an empty `OutputVar`. The first `OutputVar` is guaranteed to not be
+empty.
+
+The function will use the start date in `var.attributes["start_date"]`. The unit of time is
+expected to be second. Also, the interpolations will be inaccurate in time intervals
+outside of their respective season for the returned `OutputVar`s.
+"""
+function split_by_season_across_time(var::OutputVar)
+    _check_time_dim(var::OutputVar)
+    start_date = Dates.DateTime(var.attributes["start_date"])
+
+    seasons_across_time_dates =
+        split_by_season_across_time(time_to_date.(start_date, times(var)))
+    season_times = (
+        date_to_time.(start_date, season) for
+        season in seasons_across_time_dates
+    )
+
+    return _split_along_dim(var, time_name(var), season_times)
+end
+
 """
     check_time_dim(var::OutputVar)