+++ test/test_Var.jl
@@ -1190,6 +1190,53 @@ end
     @test_throws ErrorException ClimaAnalysis.split_by_season(var)
 end
 
+@testset "split_by_season_across_time" begin
+    lon = collect(range(-179.5, 179.5, 36))
+    lat = collect(range(-89.5, 89.5, 18))
+    time = [0.0]
+    push!(time, 2_678_400.0) # correspond to 2024-2-1
+    push!(time, 5_184_000.0) # correspond to 2024-3-1
+    push!(time, 7_862_400.0) # correspond to 2024-4-1
+    push!(time, 10_454_400.0) # correspond to 2024-5-1
+    push!(time, 13_132_800.0) # correspond to 2024-6-1
+    push!(time, 15_724_800.0) # correspond to 2024-7-1
+    push!(time, 18_403_200.0) # correspond to 2024-8-1
+    push!(time, 21_081_600.0) # correspond to 2024-9-1
+    push!(time, 36_720_000.0) # correspond to 2025-3-1
+    data = reshape(
+        1.0:1.0:(length(lat) * length(time) * length(lon)),
+        (length(lat), length(time), length(lon)),
+    )
+    dims = OrderedDict(["lat" => lat, "time" => time, "lon" => lon])
+    attribs = Dict("long_name" => "hi", "start_date" => "2024-1-1")
+    dim_attribs = OrderedDict([
+        "lat" => Dict("units" => "deg"),
+        "time" => Dict("units" => "s"),
+        "lon" => Dict("units" => "deg"),
+    ])
+    var = ClimaAnalysis.OutputVar(attribs, dims, dim_attribs, data)
+
+    seasonal_vars = ClimaAnalysis.split_by_season_across_time(var)
+    @test length(seasonal_vars) == 6
+    @test isempty(seasonal_vars[5])
+
+    # Check arrays for time is correct
+    @test seasonal_vars[1].dims["time"] == [0.0, 2_678_400.0]
+    @test seasonal_vars[2].dims["time"] ==
+          [5_184_000.0, 7_862_400.0, 10_454_400.0]
+    @test seasonal_vars[3].dims["time"] ==
+          [13_132_800.0, 15_724_800.0, 18_403_200.0]
+    @test seasonal_vars[4].dims["time"] == [21_081_600.0]
+    @test seasonal_vars[6].dims["time"] == [36_720_000.0]
+
+    # Check data itself
+    @test seasonal_vars[1].data == data[:, 1:2, :]
+    @test seasonal_vars[2].data == data[:, 3:5, :]
+    @test seasonal_vars[3].data == data[:, 6:8, :]
+    @test seasonal_vars[4].data == data[:, 9:9, :]
+    @test seasonal_vars[6].data == data[:, 10:10, :]
+end
+
 @testset "Compute bias" begin
     lon = collect(range(-179.5, 179.5, 360))
     lat = collect(range(-89.5, 89.5, 180))